+++ schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/entity/performance/PerformanceGenre.java
@@ -1,0 +1,8 @@
+package wisoft.nextframe.schedulereservationticketing.entity.performance;
+
+public enum PerformanceGenre {
+	PLAY,       // 연극
+	CONCERT,    // 콘서트
+	CLASSIC,    // 클래식
+	MUSICAL     // 뮤지컬
+}
