+++ schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/entity/performance/PerformanceType.java
@@ -1,0 +1,7 @@
+package wisoft.nextframe.schedulereservationticketing.entity.performance;
+
+public enum PerformanceType {
+	ROMANCE,    // 로맨스
+	HORROR,     // 공포
+	THRILLER    // 스릴러
+}
