+++ schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/entity/stadium/StadiumSection.java
@@ -1,0 +1,42 @@
+package wisoft.nextframe.schedulereservationticketing.entity.stadium;
+
+import java.util.UUID;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import jakarta.persistence.UniqueConstraint;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@AllArgsConstructor
+@Entity
+@Table(
+	name = "stadium_sections",
+	uniqueConstraints = {
+		@UniqueConstraint(columnNames = {"stadium_id", "section"})
+	}
+)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class StadiumSection {
+
+	@Id
+	@Column(name = "id", nullable = false)
+	private UUID id;
+
+	@ManyToOne(fetch = FetchType.LAZY, optional = false)
+	@JoinColumn(name = "stadium_id", nullable = false)
+	private Stadium stadium;
+
+	@Column(name = "section", nullable = false)
+	private String section;
+}
