+++ schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/entity/user/infra/UserEntity.java
@@ -1,0 +1,34 @@
+package wisoft.nextframe.schedulereservationticketing.entity.user.infra;
+
+import java.time.LocalDate;
+import java.util.UUID;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Entity
+@Table(name = "users")
+@NoArgsConstructor
+@AllArgsConstructor(access = AccessLevel.PROTECTED)
+@Builder
+public class UserEntity {
+
+	@Id
+	private UUID id;
+	private String name;
+
+	private LocalDate birthDate;
+
+	private String phoneNumber;
+
+	private String email;
+
+	private String faceId;
+}
