+++ schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/entity/user/infra/UserOauthEntity.java
@@ -1,0 +1,36 @@
+package wisoft.nextframe.schedulereservationticketing.entity.user.infra;
+
+import java.time.LocalDateTime;
+import java.util.UUID;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.Getter;
+
+@Getter
+@Entity
+@Table(name = "user_oauth")
+public class UserOauthEntity {
+
+	@Id
+	private UUID id;
+
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "user_id")
+	private UserEntity user;
+
+	private String provider;
+
+	private String oauthId;
+
+	@Column(columnDefinition = "text")
+	private String refreshToken;
+
+	private LocalDateTime expiresAt;
+
+}
