+++ schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/repository/performance/PerformancePricingRepository.java
@@ -1,0 +1,9 @@
+package wisoft.nextframe.schedulereservationticketing.repository.performance;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import wisoft.nextframe.schedulereservationticketing.entity.performance.PerformancePricing;
+import wisoft.nextframe.schedulereservationticketing.entity.performance.PerformancePricingId;
+
+public interface PerformancePricingRepository extends JpaRepository<PerformancePricing, PerformancePricingId> {
+}
