+++ schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/repository/performance/PerformanceRepository.java
@@ -1,0 +1,10 @@
+package wisoft.nextframe.schedulereservationticketing.repository.performance;
+
+import java.util.UUID;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import wisoft.nextframe.schedulereservationticketing.entity.performance.Performance;
+
+public interface PerformanceRepository extends JpaRepository<Performance, UUID> {
+}
