+++ schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/repository/stadium/SeatDefinitionRepository.java
@@ -1,0 +1,10 @@
+package wisoft.nextframe.schedulereservationticketing.repository.stadium;
+
+import java.util.UUID;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import wisoft.nextframe.schedulereservationticketing.entity.stadium.SeatDefinition;
+
+public interface SeatDefinitionRepository extends JpaRepository<SeatDefinition, UUID> {
+}
