--- schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/schedule/entity/performance/PerformanceGenre.java
@@ -1,8 +1,0 @@
-package wisoft.nextframe.schedulereservationticketing.schedule.entity.performance;
-
-public enum PerformanceGenre {
-	PLAY,       // 연극
-	CONCERT,    // 콘서트
-	CLASSIC,    // 클래식
-	MUSICAL     // 뮤지컬
-}
