--- schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/schedule/entity/performance/PerformanceType.java
@@ -1,7 +1,0 @@
-package wisoft.nextframe.schedulereservationticketing.schedule.entity.performance;
-
-public enum PerformanceType {
-	ROMANCE,    // 로맨스
-	HORROR,     // 공포
-	THRILLER    // 스릴러
-}
