--- schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/schedule/repository/performance/PerformancePricingRepository.java
@@ -1,9 +1,0 @@
-package wisoft.nextframe.schedulereservationticketing.schedule.repository.performance;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import wisoft.nextframe.schedulereservationticketing.schedule.entity.performance.PerformancePricing;
-import wisoft.nextframe.schedulereservationticketing.schedule.entity.performance.PerformancePricingId;
-
-public interface PerformancePricingRepository extends JpaRepository<PerformancePricing, PerformancePricingId> {
-}
