--- schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/schedule/repository/performance/PerformanceRepository.java
@@ -1,10 +1,0 @@
-package wisoft.nextframe.schedulereservationticketing.schedule.repository.performance;
-
-import java.util.UUID;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import wisoft.nextframe.schedulereservationticketing.schedule.entity.performance.Performance;
-
-public interface PerformanceRepository extends JpaRepository<Performance, UUID> {
-}
