--- schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/schedule/repository/schedule/ScheduleRepository.java
@@ -1,10 +1,0 @@
-package wisoft.nextframe.schedulereservationticketing.schedule.repository.schedule;
-
-import java.util.UUID;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import wisoft.nextframe.schedulereservationticketing.schedule.entity.schedule.Schedule;
-
-public interface ScheduleRepository extends JpaRepository<Schedule, UUID> {
-}
