+++ schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/ticketing/repository/TicketRepository.java
@@ -4,7 +4,6 @@ import java.util.Optional;
 import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 import wisoft.nextframe.schedulereservationticketing.ticketing.entity.Ticket;
