+++ schedule-reservation-ticketing/src/test/java/wisoft/nextframe/schedulereservationticketing/schedule/repository/stadium/StadiumRepositoryTest.java
@@ -11,7 +11,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
 import jakarta.transaction.Transactional;
-import wisoft.nextframe.schedulereservationticketing.schedule.entity.stadium.Stadium;
+import wisoft.nextframe.schedulereservationticketing.entity.stadium.Stadium;
+import wisoft.nextframe.schedulereservationticketing.repository.stadium.StadiumRepository;
 
 @SpringBootTest
 @Transactional
